ATR-180: moved removal of regions from property leading trivia to the code generation utils
